Moved cache utils to contrib
